extract sonar-scanner-executable params, and refactor some other stuff
